create session layer
The session layer takes care of processing the environment and arguments
to get the various nodes launched and connected to each other.  It makes
use of the spawn network layer.

This commit currently launches hostname on each node given as an
argument to the launcher executable.

There will be a followup commit that will do the getcwd and launch
itself with the correct environment so that the spawn network can start
communicating with each other.